+++ client/app/components/profile/profile.component.js
@@ -18,7 +18,9 @@ export class ProfileComponent {
     this.errors = {};
     this.user = {}; // clear it
     this.authService.getCurrentUser()
-      .then(user => angular.copy(user, this.user)); // Don't modify authService.currentUser
+      .then(user => {
+        this.user = { ...user };
+      }); // Don't modify authService.currentUser
   }
 
   clearServerError(form, fieldName) {