+++ client/app/services/teacher.service.js
@@ -8,7 +8,6 @@ export class TeacherService {
   async getTeachers() {
     const { data } = await this.$http.get('/api/user/teachers');
     this.teachers = data;
-    console.log(this.teachers);
     return this.teachers;
   }